Move ValidateUnaryContext
